split lae service + register claimed rewards
