+++ packages/api/src/validators/learnAndEarn.ts
@@ -5,10 +5,23 @@ import { defaultSchema } from './defaultSchema';
 // I wish this comes true https://github.com/hapijs/joi/issues/2864#issuecomment-1322736004
 
 // This should match Joi validations
+export type AnswerRequestType = {
+    lesson: number;
+    answers: number[];
+};
+export type StartLessonRequestType = {
+    lesson: number;
+};
+export type ListLevelsRequestType = {
+    status: string;
+    category: string;
+    level: string;
+    limit?: string;
+    offset?: string;
+};
 export type RegisterClaimRewardsRequestType = {
-    level: number;
     transactionHash: string;
-}[];
+};
 
 class LearnAndEarnValidator {
     answer = celebrate({
@@ -38,15 +51,9 @@ class LearnAndEarnValidator {
     });
 
     registerClaimRewards = celebrate({
-        body: defaultSchema
-            .array()
-            .items(
-                Joi.object({
-                    level: Joi.number().required(),
-                    transactionHash: Joi.string().required(),
-                }).required()
-            )
-            .required(),
+        body: defaultSchema.object({
+            transactionHash: Joi.string().required(),
+        }),
     });
 }
